move integration test to ops
