+++ tests/ops/test_ref_files/gdb3.mol5.XYZ
@@ -1,0 +1,10 @@
+8
+gdb3.mol5.PDB
+C          0.36100       -0.45200       -0.55100
+C         -0.71400        0.12500        0.32700
+N          0.68300        0.13300        0.74500
+H          0.44200       -1.52900       -0.61900
+H          0.67200        0.10200       -1.42800
+H         -1.36400       -0.56000        0.85700
+H         -1.14900        1.08000        0.06000
+H          1.09300        1.06300        0.63600
