+++ tests/ops/test_ref_files/h2_ref.xyz
@@ -1,0 +1,4 @@
+2
+in Angstrom
+H          0.00000        0.00000       -0.35000
+H          0.00000        0.00000        0.35000
